refactor: 拆分页面详情和搜索页面为子组件
pages/page/[wikidotId]/index.vue (2248→931行):
- PageSource: 源码查看器 + diff 对比逻辑
- PageRevisions: 修订历史列表与分页
- PageVotes: 投票列表与分页
- PageImages: 图片画廊与预览
- PageMetrics: 核心指标卡片

pages/search.vue (1953→1683行):
- SearchResults: 搜索结果展示（用户/页面/讨论区）+ IntersectionObserver 无限滚动

所有数据获取 (useAsyncData) 保留在页面级，子组件通过 props 接收数据、emit 传递事件。
不改变任何业务逻辑，纯 UI 拆分。
